+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **Changed :**
 
 - move 1-appeasing-the-masses.exs file to chapter-1 directory
+- move 2-where-did-i-put-that-value.exs file to chapter-2 directory
 
 ## [0.0.4] (September 16 2022)