+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - 5-funky-functions.exs
 - exercise-1.exs file
+- exercise-2.exs file
 
 **Changed :**